requirements.txt
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
idna==2.8
mypy==0.740
mypy-extensions==0.4.3
pycodestyle==2.5.0
PyGithub==1.44.1
PyJWT==1.7.1
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
wrapt==1.11.2
